New gio restore script
Will open a vim buffer for you to select from files that were found, so
if you have the same file trashed on multiple drives you can see it in
vim buffer and remove as necessary.
Don't know how to handle turning emojis in strings to unicode though, so
the script won't work for filenames with emojis.